+++ api/utils/index.js
@@ -1,6 +1,6 @@
 const azureUtils = require('./azureUtils');
 const cryptoUtils = require('./crypto');
-const tiktokenModels = require('./tiktokenModels');
+const { tiktokenModels, maxTokensMap } = require('./tokens');
 const migrateConversations = require('./migrateDataToFirstUser');
 const sendEmail = require('./sendEmail');
 const abortMessage = require('./abortMessage');
@@ -8,6 +8,7 @@ const abortMessage = require('./abortMessage');
 module.exports = {
   ...cryptoUtils,
   ...azureUtils,
+  maxTokensMap,
   tiktokenModels,
   migrateConversations,
   sendEmail,
