+++ api/utils/tokens.js
@@ -37,4 +37,15 @@ const models = [
   'gpt-3.5-turbo-0301'
 ];
 
-module.exports = new Set(models);
+const maxTokensMap = {
+  'gpt-4': 8191,
+  'gpt-4-0613': 8191,
+  'gpt-4-32k': 32767,
+  'gpt-4-32k-0613': 32767,
+  'gpt-3.5-turbo': 4095,
+  'gpt-3.5-turbo-0613': 4095,
+  'gpt-3.5-turbo-0301': 4095,
+  'gpt-3.5-turbo-16k': 15999,
+};
+
+module.exports = { tiktokenModels: new Set(models), maxTokensMap };
